+++ jest.config.js
@@ -12,5 +12,6 @@ module.exports = {
   moduleNameMapper: {
     '^@signalman/otel(|/.*)$': '<rootDir>/libs/otel/src/$1',
     '^@signalman/propagation(|/.*)$': '<rootDir>/libs/propagation/src/$1',
+    '^@signalman/logging(|/.*)$': '<rootDir>/libs/logging/src/$1',
   },
 };